requirements.txt
python-dotenv~=1.0.1
pandas~=2.2.2
selenium~=4.31.0
bs4~=0.0.2
beautifulsoup4~=4.12.3
undetected-chromedriver~=3.5.5
urllib3~=1.26.9
scikit-learn~=1.4.2
matplotlib~=3.5.2
numpy~=1.24.4
plotly~=5.24.1